formatter

+++ external/homiphily.py
@@ -1,8 +1,10 @@
 """Sourced from https://github.com/CUAI/Non-Homophily-Large-Scale/blob/master/homophily.py"""
 
-from torch_geometric.utils import remove_self_loops
 import torch
+from torch_geometric.utils import remove_self_loops
+
 from external.torch_scatter import scatter_add
+
 
 def compat_matrix_edge_idx(edge_idx, labels):
     """
@@ -13,22 +15,26 @@ def compat_matrix_edge_idx(edge_idx, labels):
      treats negative labels as unlabeled
      """
     edge_index = remove_self_loops(edge_idx)[0]
-    src_node, targ_node = edge_index[0,:], edge_index[1,:]
+    src_node, targ_node = edge_index[0, :], edge_index[1, :]
     labeled_nodes = (labels[src_node] >= 0) * (labels[targ_node] >= 0)
     label = labels.squeeze()
     c = label.max()+1
-    H = torch.zeros((c,c)).to(edge_index.device)
+    H = torch.zeros((c, c)).to(edge_index.device)
     src_label = label[src_node[labeled_nodes]]
     targ_label = label[targ_node[labeled_nodes]]
-    label_idx = torch.cat((src_label.unsqueeze(0), targ_label.unsqueeze(0)), axis=0)
+    label_idx = torch.cat(
+        (src_label.unsqueeze(0), targ_label.unsqueeze(0)), axis=0)
     for k in range(c):
         sum_idx = torch.where(src_label == k)[0]
         add_idx = targ_label[sum_idx]
-        scatter_add(torch.ones_like(add_idx).to(H.dtype), add_idx, out=H[k,:], dim=-1)
+        scatter_add(torch.ones_like(add_idx).to(
+            H.dtype), add_idx, out=H[k, :], dim=-1)
     H = H / torch.sum(H, axis=1, keepdims=True)
     return H
 
 # ADDING THE FOLLOWING HELPER BASED ON OUR NEED TO PROCESS MANY GRAPHS
+
+
 def padded_counts_and_props(label):
     """
     Returns:
@@ -47,7 +53,8 @@ def padded_counts_and_props(label):
                 0)
 
     labeled = label[mask]
-    present, counts = torch.unique(labeled, return_counts=True)  # sorted by value
+    present, counts = torch.unique(
+        labeled, return_counts=True)  # sorted by value
     c = int(labeled.max().item()) + 1
 
     counts_full = torch.zeros(c, dtype=counts.dtype, device=label.device)
@@ -69,15 +76,14 @@ def our_measure(edge_index, label):
     c = label.max()+1
     H = compat_matrix_edge_idx(edge_index, label)
     nonzero_label = label[label >= 0]
-    #counts = nonzero_label.unique(return_counts=True)[1]
-    #proportions = counts.float() / nonzero_label.shape[0]
+    # counts = nonzero_label.unique(return_counts=True)[1]
+    # proportions = counts.float() / nonzero_label.shape[0]
     counts, proportions, _ = padded_counts_and_props(label)
     val = 0
     for k in range(c):
-        class_add = torch.clamp(H[k,k] - proportions[k], min=0)
+        class_add = torch.clamp(H[k, k] - proportions[k], min=0)
         if not torch.isnan(class_add):
             # only add if not nan
             val += class_add
     val /= c-1
     return val
-
